Remove cdbapi.
In order to get cdbapi to work in our dev env, we would have to re
compile PHP with the oracle8 drivers.  Since we aren't planning to keep
using Oracle it doesn't make sense to set it up.  Instead we will dump
the data we need using Dev into structured flat files that we can use to
import from.