Add tokens and addressfiled_tokens modules. #22
These modules are required to be enabled while building the view
if you want to access the individual fields of an address of a
referenced entity. But, they do not need to be enabled for the view to work.
So, they are in the repo so we know we used then, but will not be enabled by default.